+++ packages/proto-gen-dts/src/index.ts
@@ -1,4 +1,7 @@
 import protoGenDTS from './gen-dts';
+
+export { default as writeDTS, parseNameSpace, protoTypeToJSType } from './write-dts';
+export { default as writeReference } from './write-reference';
 export * from './type.d';
 
 export default protoGenDTS;